requirements.txt
Flask==1.1.2
numpy==1.23.5
pandas==2.0.3
scikit_learn==1.3.1
umap_learn==0.5.4
